Successfully made a second page for the website
